requirements.txt
# MLB Betting System Requirements
# Core dependencies for the betting system

# API and HTTP requests
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
networkx>=3.2.0

# Machine learning and AI
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Database and storage
supabase>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging and monitoring
psutil>=5.9.0
structlog>=23.2.0

# Date and time handling
python-dateutil>=2.8.0
pytz>=2023.3

# JSON and data serialization
orjson>=3.9.0
ujson>=5.8.0

# Async support
asyncio-mqtt>=0.16.0
aioredis>=2.0.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Utilities
click>=8.1.0
rich>=13.6.0
tqdm>=4.66.0
tabulate>=0.9.0

# Web framework (if needed)
fastapi>=0.104.0
uvicorn>=0.24.0

# Background tasks
celery>=5.3.0
redis>=5.0.0

# File handling
openpyxl>=3.1.0
xlrd>=2.0.0

# Image processing (for thumbnails)
Pillow>=10.0.0

# Security
cryptography>=41.0.0
passlib>=1.7.0

# Performance monitoring
memory-profiler>=0.61.0
line-profiler>=4.1.0

# Development tools 
pre-commit>=3.5.0
isort>=5.12.0
joblib
xgboost
scikit-learn
openpyxl
lxml
